sweet alert add

+++ src/Pages/MyServices.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../Provider/AuthProvider';
 import { Link } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const MyServices = () => {
 
@@ -19,14 +20,40 @@ const MyServices = () => {
 
 
           const handleDelete = (id) => {
-                axios.delete(`http://localhost:3000/delete/${id}`)
-                 .then(res => {
-                   console.log(res.data)
-                   const filterData = myServices.filter(service => service?._id !== id )
-                   setMyServices(filterData);
 
-                 })
-                 .catch(err => console.log(err))
+                Swal.fire({
+                    title: "Are you sure?",
+                    text: "You won't be able to revert this!",
+                    icon: "warning",
+                    showCancelButton: true,
+                    confirmButtonColor: "#3085d6",
+                    cancelButtonColor: "#d33",
+                    confirmButtonText: "Yes, delete it!"
+                    }).then((result) => {
+                            
+                        if (result.isConfirmed) {
+
+                         axios.delete(`http://localhost:3000/delete/${id}`)
+                            .then(res => {
+                            console.log(res.data)
+
+                            if(res.data.acknowledged) {
+                               const filterData = myServices.filter(service => service?._id !== id )
+                               setMyServices(filterData);
+                                Swal.fire({
+                                    title: "Deleted!",
+                                    text: "Your service has been deleted.",
+                                    icon: "success"
+                                });
+                           } 
+                            
+                            })
+                            .catch(err => console.log(err))
+
+                }
+                });
+
+               
           }
 
     return (
